// FamilyClub.React/famiy.club.react/lib/api/generated/models/LanguageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LanguageDto
 */
export interface LanguageDto {
    /**
     * 
     * @type {number}
     * @memberof LanguageDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof LanguageDto
     */
    languageName?: string | null;
}

/**
 * Check if a given object implements the LanguageDto interface.
 */
export function instanceOfLanguageDto(value: object): value is LanguageDto {
    return true;
}

export function LanguageDtoFromJSON(json: any): LanguageDto {
    return LanguageDtoFromJSONTyped(json, false);
}

export function LanguageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): LanguageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'languageName': json['languageName'] == null ? undefined : json['languageName'],
    };
}

export function LanguageDtoToJSON(json: any): LanguageDto {
    return LanguageDtoToJSONTyped(json, false);
}

export function LanguageDtoToJSONTyped(value?: LanguageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'languageName': value['languageName'],
    };
}